Move Adapter to storage_v2 Package
