+++ storage_v2/factoryadapter/factory.go
