+++ storage_v2/factoryadapter/writer.go
