+++ storage_v2/factoryadapter/writer_test.go
